adding recording of rx stats, sadly using stats class tho

+++ drivers/net/ethernet/mellanox/mlx5/core/intlog.c
@@ -400,7 +400,7 @@ void record_tx_poll_info(u16 npkts, u32 nbytes) { //gonna have to cast types
 	per_irq_stats.tx_npkts += (unsigned int)npkts;
 }
 
-void record_rx_poll_info(u16 npkts, u32 nbytes) {
+void record_rx_poll_info(uint64_t npkts, uint64_t nbytes) {
 	per_irq_stats.rx_nbytes += (unsigned int)nbytes;
 	per_irq_stats.rx_npkts += (unsigned int)npkts;
 }
